backend/requirements.txt
fastapi
uvicorn
pydantic
motor[srv]
python-dotenv
certifi
dnspython
email-validator
